refactor: extract RepoSelect
