+++ components/editor/modules/meta/ui.js
@@ -2,15 +2,14 @@ import React from 'react'
 import { css } from 'glamor'
 import { Map, Set } from 'immutable'
 
-import { Interaction, A, Dropdown, Field, Label, colors } from '@project-r/styleguide'
+import { Interaction, Dropdown, Field, Label, colors } from '@project-r/styleguide'
 
-import { GITHUB_ORG } from '../../../../lib/settings'
 import withT from '../../../../lib/withT'
 import MetaForm from '../../utils/MetaForm'
-import RepoSearch from '../../utils/RepoSearch'
 import SlugField from '../../utils/SlugField'
 import FBPreview from './FBPreview'
 import TwitterPreview from './TwitterPreview'
+import RepoSelect from './RepoSelect'
 import UIForm from '../../UIForm'
 
 const styles = {
@@ -84,6 +83,7 @@ const MetaData = ({value, editor, additionalFields = [], customFields = [], teas
         </Interaction.H2>
         <br />
         <SlugField
+          black
           label={t(`metaData/field/slug`, undefined, 'slug')}
           value={node.data.get('slug')}
           onChange={onInputChange('slug')}
@@ -103,60 +103,36 @@ const MetaData = ({value, editor, additionalFields = [], customFields = [], teas
                 onChange={item => onChange(undefined, item.value)} />
             }
             if (customField.ref === 'repo') {
-              const onRefChange = item => {
-                if (customField.key === 'format') {
-                  editor.change(change => {
-                    change
-                      .setNodeByKey(node.key, {
-                        data: item
-                          ? node.data.set('format', `https://github.com/${item.value.id}`)
-                          : node.data.remove('format')
-                      })
-                    let titleNode = change.value.document
-                      .findDescendant(node => node.type === 'TITLE')
-                    if (titleNode) {
-                      const format = item
-                        ? item.value.latestCommit.document
-                        : undefined
-                      change.setNodeByKey(titleNode.key, {
-                        data: {format}
-                      })
-                      titleNode.nodes.forEach(node => {
-                        change.setNodeByKey(node.key, {
+              return <RepoSelect key={customField.key}
+                label={label}
+                value={value}
+                onChange={customField.key === 'format'
+                  ? (_, __, item) => {
+                    editor.change(change => {
+                      change
+                        .setNodeByKey(node.key, {
+                          data: item
+                            ? node.data.set('format', `https://github.com/${item.value.id}`)
+                            : node.data.remove('format')
+                        })
+                      let titleNode = change.value.document
+                        .findDescendant(node => node.type === 'TITLE')
+                      if (titleNode) {
+                        const format = item
+                          ? item.value.latestCommit.document
+                          : undefined
+                        change.setNodeByKey(titleNode.key, {
                           data: {format}
                         })
-                      })
-                    }
-                  })
-                  return
-                }
-                onChange(
-                  undefined,
-                  item
-                    ? `https://github.com/${item.value.id}`
-                    : null
-                )
-              }
-              if (value) {
-                return (
-                  <div style={{height: 60, marginBottom: 12, borderBottom: '1px solid #000'}}>
-                    <Label style={{color: '#000'}}>{label}</Label><br />
-                    <Interaction.P>
-                      {String(value)
-                        .replace('https://github.com/', '')
-                        .replace(`${GITHUB_ORG}/`, '')}
-                      {' '}
-                      <A href='#remove' onClick={() => {
-                        onRefChange(null)
-                      }}>x</A>
-                    </Interaction.P>
-                  </div>
-                )
-              }
-              return <RepoSearch key={customField.key}
-                label={label}
-                onChange={onRefChange}
-              />
+                        titleNode.nodes.forEach(node => {
+                          change.setNodeByKey(node.key, {
+                            data: {format}
+                          })
+                        })
+                      }
+                    })
+                  }
+                  : onChange} />
             }
             return <Field key={customField.key}
               black
